LayoutTests/http/tests/websocket/tests/hybi/long-control-frame-expected.txt

CONSOLE MESSAGE: Received control frame having too long payload: 126 bytes
Test whether WebSocket rejects control frames longer than 125 bytes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
